Unify logic to render embedded Morphs inside of Text
This now uses the same logic as every other morph for the most part.
The actually embedded morph gets rendered inside of the line by `renderMorphInLine`. After that, the usual loop kicks in. 

Embedded morphs are part of the normal submorph array. To identify which submorphs are embedded and which ones are "normal", we use the `embeddedMorphMap` of the Text, which gets populated upon replacement of the document.